fwefw

+++ src/main/kotlin/org/mmarket/clans/system/model/ClanSlots.kt
@@ -2,14 +2,14 @@ package org.mmarket.clans.system.model
 
 import org.mmarket.clans.files.Settings
 
-enum class ClanSlots(val plus: Int, val cost: Long) {
-    INITIAL(Settings.long("slots.initial.plus").toInt(), Settings.long("slots.initial.cost")),
-    FIRST(Settings.long("slots.first.plus").toInt(), Settings.long("slots.first.cost")),
-    SECOND(Settings.long("slots.second.plus").toInt(), Settings.long("slots.second.cost")),
-    THIRD(Settings.long("slots.third.plus").toInt(), Settings.long("slots.third.cost")),
-    FOURTH(Settings.long("slots.fourth.plus").toInt(), Settings.long("slots.fourth.cost")),
-    FIFTH(Settings.long("slots.fifth.plus").toInt(), Settings.long("slots.fifth.cost")),
-    SIXTH(Settings.long("slots.sixth.plus").toInt(), Settings.long("slots.sixth.cost"));
+enum class ClanSlots(val plus: Int, val cost: Double) {
+    INITIAL(Settings.long("slots.initial.plus").toInt(), Settings.double("slots.initial.cost")),
+    FIRST(Settings.long("slots.first.plus").toInt(), Settings.double("slots.first.cost")),
+    SECOND(Settings.long("slots.second.plus").toInt(), Settings.double("slots.second.cost")),
+    THIRD(Settings.long("slots.third.plus").toInt(), Settings.double("slots.third.cost")),
+    FOURTH(Settings.long("slots.fourth.plus").toInt(), Settings.double("slots.fourth.cost")),
+    FIFTH(Settings.long("slots.fifth.plus").toInt(), Settings.double("slots.fifth.cost")),
+    SIXTH(Settings.long("slots.sixth.plus").toInt(), Settings.double("slots.sixth.cost"));
 
     fun next(): ClanSlots {
         return when (this) {
